+++ update_launch_other.go
@@ -1,0 +1,9 @@
+//go:build !windows
+
+package main
+
+import "errors"
+
+func startDetachedUpdateHelper(string, ...string) error {
+	return errors.New("automatic updates are only available on Windows")
+}